+++ changelog.md
@@ -7,6 +7,7 @@
 * Added remote macro download counts.
 * Added the default cloud macro server connection.
 * Fixed remote macro downloads for filenames with spaces.
+* Fixed remote macro Load only downloading instead of loading the macro.
 * Updated to Geode 5.7.1 and Geometry Dash 2.2081.
 * Added Android64 package support.
 * Fixed spider being buggy if using Input Fixes.